% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBEPS-package.R
\docType{package}
\name{rBEPS-package}
\alias{rBEPS}
\alias{rBEPS-package}
\title{rBEPS: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
