% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readcoef_}
\alias{readcoef_}
\title{Real soil coefficients}
\usage{
readcoef_(lc, stxt)
}
\arguments{
\item{lc}{land cover type
1-ENF 2-DNF 6-DBF 9-EBF 13-Shrub 40-C4 Plants default:Others}

\item{stxt}{soil texture, 1-11}
}
\description{
Real soil coefficients
}
\examples{
readcoef_(1, 1)
}
