% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBEPS-package.R
\docType{package}
\name{rBEPS-package}
\alias{rBEPS}
\alias{rBEPS-package}
\title{rBEPS: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\keyword{internal}
